LRQA-41845 Make Label a static inner class to match how Comment is implemented
